🚚 Rename: 폴더 이름 형식을 통일
