+++ src/components/layout/head-nav/HeadNav.tsx
@@ -2,7 +2,7 @@ import Link from "next/link";
 import Image from "next/image";
 import * as S from "./HeadNav.style";
 
-import HeadNavProfile from "./Profile/Profile";
+import HeadNavProfile from "./Profile";
 
 export default function HeadNav({ isSticky = true }) {
   return (
